% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_dtm.R
\name{cast_dtm}
\alias{cast_dtm}
\title{Casting a data frame to
a DocumentTermMatrix}
\usage{
cast_dtm(data, document, term, value, weighting = tm::weightTf, ...)
}
\arguments{
\item{data}{Table with one-term-per-document-per-row}

\item{document}{Column containing document IDs as string or symbol}

\item{term}{Column containing terms as string or symbol}

\item{value}{Column containing values as string or symbol}

\item{weighting}{The weighting function for the DTM/TDM
(default is term-frequency, effectively unweighted)}

\item{...}{Extra arguments passed on to
\code{\link{sparseMatrix}}}
}
\description{
This turns a "tidy" one-term-per-document-per-row data frame into a
DocumentTermMatrix from the tm package. This function supports non-standard
evaluation through the tidyeval framework. Groups are ignored.
}
\details{
The arguments \code{term}, \code{document}, and \code{value}
are passed by expression and support \link[rlang]{quasiquotation};
you can unquote strings and symbols.
}
