% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{LDA}
\alias{LDA}
\alias{match_terms}
\title{Latent Dirichlet Allocation (LDA)}
\usage{
(x, k, method = "VEM", control = NULL, model = NULL, ...)
}
\arguments{
\item{x}{Any object coercible to such simple triplet matrices (with count entries),}

\item{model}{Object of class "LDA" for initialization}

\item{k}{Integer; number of topics}

\item{method}{The method to be used for fitting the model; either "VEM" or "Gibbs".}

\item{control}{A named listed of the control parameters for estimation}

\item{...}{Optional arguments. For method = "Gibbs" an additional argument seedwords can be specified as a matrix or an object of class "simple_triplet_matric".}
}
\description{
Estimate a LDA model using either a VEN algorithm or Gibbs sampling.
}
\examples{
LDA(x, k, method = "VEM", control = NULL, model = NULL)
}
\keyword{LDA}
