% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets.R
\docType{data}
\name{tweets.2020}
\alias{tweets.2020}
\title{Data from  ballot harvesting, Trump's taxes, and SCOTUS, scraped from Twitter in 2020.}
\format{
A data frame with 91 columns and 46,477 rows
}
\usage{
data(tweets.2020)
}
\description{
Data scraped from Twitter in 2020. Contains tweets related to ballot harvesting, Trump's taxes, and SCOTUS.
}
\keyword{datasets}
